Map.java Updated
My comment before was incorrect it was 18.5 not 24.5 anyway checkData(){/*gets all points in a four step radius and tests them for loadedness. This could be one of two things ether the map is hidden by a screen and the data is there or the data has not been generated yet  no need for a screen. i like the second but that might mean we need a list of places we need and add them in Random order to grid.  If the location has been loaded it should do nothing. in this aspect having a screen would be easier because we would just show what is under screen but where is the fun in that*/}